Add feature vendored-protoc to use protobuf-src crate.
With this feature, which defaults to being enabled, the protobuf compiler is
built from C sources at build time, instead of using a locally installed protoc
compiler. This is intended to make it easier for users of our crate (it turns
out to be very difficult to ensure that protoc is installed on build machines
when using Github actions such as the upload-rust-binary-action, which depends
on cross-rs running in a container).